requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
solace-pubsubplus==1.1.0
urllib3==1.26.4
